% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-extract-L1trend.R
\name{l1trend}
\alias{l1trend}
\title{wrapper function for fitting L1 trend estimator l1tf from package l1tf}
\usage{
l1trend(y.v, prop, sens = 100, max.length = 5, max.prop.na = 0.2)
}
\arguments{
\item{y.v}{A numeric vector}

\item{prop}{A scalar showing the proportion of NAs that are allowable in y.v}

\item{sens}{A sensitivity parameter that is involved in identifying break-points, defaults to 100}

\item{max.length}{the maximum allowable segment length, defaults to 5}

\item{max.prop.na}{the maximum allowable proportion of NAs within the input vector}
}
\value{
An l1tf_obj with slots summary.df, series.df and prop.na
}
\description{
\code{l1trend} returns an object of class l1tf_obj
}
\details{
This is a wrapper function for l1tf from the l1tf package.  The l1tf function fits an L1 regularised
piecewise linear function to a time series data set.  The wrapper function processes the vector returned
by l1tf and returns an S4 object with three slots, these being: summary.df, series.df and prop.na.
summary.df provides a summary of the linear functions and change-points.
series.df provides an output of the time-series with imputed NA values and piecewise linear trend.
prop.na provides the number of NAs as a proportion of the series length.
There is an associated summary and plot method fo rthe l1tf_obj class.
}
\examples{
\dontrun{data(somestocks)
data.sub <- somestocks[,1]
trend.est <- l1trend(y.v = data.sub,prop=0.2,sens=100,max.length=5,max.prop.na=0.2)
summary(trend.est)
plot(trend.est)
}

}

