% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{somestocks}
\alias{somestocks}
\title{Stock prices from the ASX}
\format{A data frame with 350 rows (trading days) and 20 columns (stocks)}
\usage{
somestocks
}
\description{
A dataset containing stock prices of a set of stocks listed on the ASX.  There are 20 stocks in all and their price is
given for a period of 350 trading days
}
\keyword{datasets}

