% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\docType{methods}
\name{summary,l1tf_obj-method}
\alias{summary,l1tf_obj-method}
\title{A summary method for the l1tf output}
\usage{
\S4method{summary}{l1tf_obj}(object, ...)
}
\arguments{
\item{object}{An object of class l1tf_obj}
}
\description{
\code{summary} returns a summary of the l1tf output
}

