% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{l1tf_obj-class}
\alias{l1tf_obj-class}
\title{An S4 class to represent l1tf output}
\description{
An S4 class to represent l1tf output
}
\section{Slots}{

\describe{
\item{\code{summary.df}}{A numeric dataframe}

\item{\code{series.df}}{A numeric dataframe}

\item{\code{prop.na}}{A length-one numeric vector}
}}

