% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\docType{methods}
\name{plot,l1tf_obj,missing-method}
\alias{plot,l1tf_obj,missing-method}
\title{A plot method for the l1tf output}
\usage{
\S4method{plot}{l1tf_obj,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class l1tf_obj}
}
\description{
\code{plot} returns a plot of the l1tf output
}

